NEW (#8): Ability to apply grid field capability to pages via DataExtension.
