+++ README.md
@@ -19,6 +19,7 @@ This module is meant as base classes, it can be used on its own but usually you 
 *These will get auto-installed when using composer:*
 
 * SilverStripe 3.0 or newer
+* PHP 5.4 or newer
 * [silverstripe-excludechildren module to hide pages from the sitetree](https://github.com/micschk/silverstripe-excludechildren)
 * [silverstripe-gridfieldsitetreebuttons to manage SiteTree items in a gridfield](https://github.com/micschk/silverstripe-gridfieldsitetreebuttons)
